# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Register a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Logout (revoke refresh session)",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user (requires JWT)",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/password/request": {
      "post": {
        "operationId": "AuthController_requestPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Request password reset (no account enumeration)",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/password/confirm": {
      "post": {
        "operationId": "AuthController_confirmPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Confirm password reset with token",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/tos/{version}": {
      "get": {
        "operationId": "TosController_getVersion",
        "parameters": [
          {
            "name": "version",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a specific Terms of Service version",
        "tags": [
          "tos"
        ]
      }
    },
    "/v1/tos": {
      "post": {
        "operationId": "TosController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTosDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new Terms of Service version (admin only)",
        "tags": [
          "tos"
        ]
      }
    },
...
    "/v1/admin/tags": {
      "post": {
        "operationId": "TagsController_createTag",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a tag (admin only)",
        "tags": [
          "tags"
        ]
      },
      "get": {
        "operationId": "TagsController_listTags",
        "parameters": [
          {
            "name": "includeInactive",
            "required": false,
            "in": "query",
            "description": "Include inactive tags",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by name or description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all tags with filters (admin only)",
        "tags": [
          "tags"
        ]
      }
    },
    "/v1/admin/tags/{id}": {
      "get": {
        "operationId": "TagsController_getTagById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get tag by ID (admin only)",
        "tags": [
          "tags"
        ]
      },
      "patch": {
        "operationId": "TagsController_updateTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update tag (admin only)",
        "tags": [
          "tags"
        ]
      },
      "delete": {
        "operationId": "TagsController_deleteTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete/disable tag (admin only)",
        "tags": [
          "tags"
        ]
      }
    },
...
    "/v1/api/upload-url": {
      "post": {
        "description": "Returns a signed URL for uploading files directly to R2 storage, plus a public URL for storing in the database",
        "operationId": "UploadController_generateUploadUrl",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateUploadUrlDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Presigned URLs generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "signedUrl": "https://your-bucket.r2.cloudflarestorage.com/...",
                    "publicUrl": "https://cdn.yourdomain.com/image/123456.jpg",
                    "key": "image/123456-abc123.jpg"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate presigned URL for direct cloud upload",
        "tags": [
          "Upload"
        ]
      }
    }
  },
  "info": {
    "title": "Lendly API",
    "description": "Lendly backend API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "phone": {
            "type": "string",
            "example": "+15551234567"
          },
          "password": {
            "type": "string",
            "example": "securePassword123",
            "minLength": 6
          },
          "acceptTerms": {
            "type": "boolean",
            "example": true
          },
          "acceptPrivacy": {
            "type": "boolean",
            "example": true
          },
          "tosLocale": {
            "type": "string",
            "example": "en"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "phone",
          "password",
          "acceptTerms",
          "acceptPrivacy"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "securePassword123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token to revoke"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "RequestPasswordResetDto": {
        "type": "object",
        "properties": {}
      },
      "ConfirmPasswordResetDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateTosDto": {
        "type": "object",
        "properties": {
          "locale": {
            "type": "string",
            "example": "en",
            "description": "Locale code for the TOS document"
          },
          "content": {
            "type": "string",
            "example": "# Terms of Service\nThis is the legal text...",
            "description": "Markdown content of the Terms of Service"
          },
          "requiresReacceptance": {
            "type": "boolean",
            "example": false,
            "description": "Whether this version requires users to re-accept in the future"
          }
        },
        "required": [
          "locale",
          "content"
        ]
      },
...
      "CreateTagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Outdoor",
            "minLength": 2,
            "maxLength": 50
          },
          "description": {
            "type": "string",
            "example": "Items mostly used outside"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "default": true
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateTagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Outdoor"
          },
          "description": {
            "type": "string",
            "example": "Items mostly used outside"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "GenerateUploadUrlDto": {
        "type": "object",
        "properties": {
          "fileType": {
            "type": "string",
            "example": "image",
            "enum": [
              "image",
              "video",
              "document"
            ],
            "description": "Type of file being uploaded"
          },
          "extension": {
            "type": "string",
            "example": "jpg",
            "description": "File extension (without dot)"
          },
          "folder": {
            "type": "string",
            "example": "listing-photos",
            "description": "Optional subfolder path"
          }
        },
        "required": [
          "fileType"
        ]
      }
    }
  }
}